# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type AuthProfileRole = typeof AuthProfileRole[keyof typeof AuthProfileRole];


export const AuthProfileRole = {
  patient: 'patient',
  reception: 'reception',
  nursing: 'nursing',
  doctor: 'doctor',
  coordination: 'coordination',
} as const;

export interface AuthProfile {
  id: number;
  role: AuthProfileRole;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  patientId: number | null;
  /** @nullable */
  doctorId: number | null;
  createdAt: string;
}

export interface LinkPatientInput {
  /** @minLength 3 */
  medicalRecordNumber: string;
  /** @minLength 3 */
  phone: string;
}

export type PatientSex = typeof PatientSex[keyof typeof PatientSex];


export const PatientSex = {
  female: 'female',
  male: 'male',
  other: 'other',
} as const;

export interface Patient {
  id: number;
  medicalRecordNumber: string;
  name: string;
  sex: PatientSex;
  birthDate: string;
  phone: string;
  /** @nullable */
  neighborhood?: string | null;
  createdAt: string;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  scheduled: 'scheduled',
  confirmed: 'confirmed',
  waiting: 'waiting',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export type AppointmentType = typeof AppointmentType[keyof typeof AppointmentType];


export const AppointmentType = {
  first_visit: 'first_visit',
  follow_up: 'follow_up',
  return: 'return',
} as const;

export interface Appointment {
  id: number;
  patientId: number;
  patientName: string;
  medicalRecordNumber: string;
  departmentId: number;
  departmentName: string;
  doctorId: number;
  doctorName: string;
  date: string;
  time: string;
  status: AppointmentStatus;
  type: AppointmentType;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PatientPortal {
  profile: AuthProfile;
  patient: Patient | null;
  appointments: Appointment[];
}

export type PatientMessageSenderRole = typeof PatientMessageSenderRole[keyof typeof PatientMessageSenderRole];


export const PatientMessageSenderRole = {
  patient: 'patient',
  doctor: 'doctor',
  nursing: 'nursing',
  coordination: 'coordination',
} as const;

export interface PatientMessage {
  id: number;
  patientId: number;
  /** @nullable */
  doctorId: number | null;
  senderRole: PatientMessageSenderRole;
  body: string;
  createdAt: string;
  /** @nullable */
  readAt?: string | null;
}

export interface PatientMessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}

export interface HealthStatus {
  status: string;
}

export interface Department {
  id: number;
  name: string;
  color: string;
  active: boolean;
}

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  departmentId: number;
  initials: string;
  active: boolean;
}

export type PatientInputSex = typeof PatientInputSex[keyof typeof PatientInputSex];


export const PatientInputSex = {
  female: 'female',
  male: 'male',
  other: 'other',
} as const;

export interface PatientInput {
  /** @minLength 2 */
  name: string;
  sex: PatientInputSex;
  birthDate: string;
  phone: string;
  neighborhood?: string;
}

export type PatientDetail = Patient & {
  appointments: Appointment[];
};

export type AppointmentInputType = typeof AppointmentInputType[keyof typeof AppointmentInputType];


export const AppointmentInputType = {
  first_visit: 'first_visit',
  follow_up: 'follow_up',
  return: 'return',
} as const;

export interface AppointmentInput {
  patientId: number;
  departmentId: number;
  doctorId: number;
  date: string;
  time: string;
  type: AppointmentInputType;
  notes?: string;
}

export type AppointmentUpdateStatus = typeof AppointmentUpdateStatus[keyof typeof AppointmentUpdateStatus];


export const AppointmentUpdateStatus = {
  scheduled: 'scheduled',
  confirmed: 'confirmed',
  waiting: 'waiting',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export interface AppointmentUpdate {
  date?: string;
  time?: string;
  doctorId?: number;
  status?: AppointmentUpdateStatus;
  notes?: string;
}

export interface DepartmentSummary {
  departmentId: number;
  departmentName: string;
  total: number;
  completed: number;
  waiting: number;
}

export interface DashboardSummary {
  date: string;
  totalAppointments: number;
  completedAppointments: number;
  waitingPatients: number;
  inProgressAppointments: number;
  cancelledAppointments: number;
  averageWaitMinutes?: number;
  departments: DepartmentSummary[];
  nextAppointment: Appointment | null;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  appointment_created: 'appointment_created',
  patient_registered: 'patient_registered',
  status_updated: 'status_updated',
  appointment_completed: 'appointment_completed',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  /** @nullable */
  actor?: string | null;
  timestamp: string;
}

export type ListAppointmentsParams = {
date?: string;
status?: ListAppointmentsStatus;
departmentId?: number;
search?: string;
};

export type ListAppointmentsStatus = typeof ListAppointmentsStatus[keyof typeof ListAppointmentsStatus];


export const ListAppointmentsStatus = {
  scheduled: 'scheduled',
  confirmed: 'confirmed',
  waiting: 'waiting',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
  no_show: 'no_show',
} as const;

export type ListPatientsParams = {
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListDoctorsParams = {
departmentId?: number;
};

export type ListActivityParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};