# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Operational summaries
  - name: appointments
    description: Consultation scheduling and queue
  - name: patients
    description: Patient records
  - name: directory
    description: Hospital departments and clinicians
  - name: auth
    description: Authenticated user profiles and patient portal
paths:
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get the authenticated user's hospital profile
      responses:
        "200":
          description: Authenticated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthProfile"
        "401":
          description: Authentication required
  /auth/link-patient:
    post:
      operationId: linkPatient
      tags: [auth]
      summary: Link an authenticated account to a patient record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkPatientInput"
      responses:
        "200":
          description: Profile linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthProfile"
        "404":
          description: Matching patient record not found
  /auth/patient-portal:
    get:
      operationId: getPatientPortal
      tags: [auth]
      summary: Get the authenticated patient's private portal data
      responses:
        "200":
          description: Patient portal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientPortal"
  /auth/messages:
    get:
      operationId: listPatientMessages
      tags: [auth]
      summary: List secure messages for the authenticated patient
      responses:
        "200":
          description: Patient messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientMessage"
    post:
      operationId: sendPatientMessage
      tags: [auth]
      summary: Send a secure message to the patient's attending doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientMessage"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get today's consultation operations summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List consultation appointments
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, waiting, in_progress, completed, cancelled, no_show]
        - name: departmentId
          in: query
          schema: { type: number }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Schedule a consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update consultation status or details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: Search patients
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: number, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Patient registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient details and consultation history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Patient detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDetail"
        "404":
          description: Patient not found
  /departments:
    get:
      operationId: listDepartments
      tags: [directory]
      summary: List hospital consultation departments
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
  /doctors:
    get:
      operationId: listDoctors
      tags: [directory]
      summary: List clinicians
      parameters:
        - name: departmentId
          in: query
          schema: { type: number }
      responses:
        "200":
          description: Clinicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /activity:
    get:
      operationId: listActivity
      tags: [dashboard]
      summary: List recent operational activity
      parameters:
        - name: limit
          in: query
          schema: { type: number, minimum: 1, maximum: 50, default: 10 }
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
    AuthProfile:
      type: object
      required: [id, role, patientId, doctorId, createdAt]
      properties:
        id: { type: number }
        role: { type: string, enum: [patient, reception, nursing, doctor, coordination] }
        displayName: { type: ["string", "null"] }
        patientId: { type: ["number", "null"] }
        doctorId: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
    LinkPatientInput:
      type: object
      required: [medicalRecordNumber, phone]
      properties:
        medicalRecordNumber: { type: string, minLength: 3 }
        phone: { type: string, minLength: 3 }
    PatientPortal:
      type: object
      required: [profile, patient, appointments]
      properties:
        profile:
          $ref: "#/components/schemas/AuthProfile"
        patient:
          anyOf:
            - $ref: "#/components/schemas/Patient"
            - type: "null"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
    PatientMessage:
      type: object
      required: [id, patientId, doctorId, senderRole, body, createdAt]
      properties:
        id: { type: number }
        patientId: { type: number }
        doctorId: { type: ["number", "null"] }
        senderRole: { type: string, enum: [patient, doctor, nursing, coordination] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        readAt: { type: ["string", "null"], format: date-time }
    PatientMessageInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 2000 }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Department:
      type: object
      required: [id, name, color, active]
      properties:
        id: { type: number }
        name: { type: string }
        color: { type: string }
        active: { type: boolean }
    Doctor:
      type: object
      required: [id, name, specialty, departmentId, initials, active]
      properties:
        id: { type: number }
        name: { type: string }
        specialty: { type: string }
        departmentId: { type: number }
        initials: { type: string }
        active: { type: boolean }
    Patient:
      type: object
      required: [id, medicalRecordNumber, name, sex, birthDate, phone, createdAt]
      properties:
        id: { type: number }
        medicalRecordNumber: { type: string }
        name: { type: string }
        sex:
          type: string
          enum: [female, male, other]
        birthDate: { type: string, format: date }
        phone: { type: string }
        neighborhood: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    PatientInput:
      type: object
      required: [name, sex, birthDate, phone]
      properties:
        name: { type: string, minLength: 2 }
        sex:
          type: string
          enum: [female, male, other]
        birthDate: { type: string, format: date }
        phone: { type: string }
        neighborhood: { type: string }
    PatientDetail:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          required: [appointments]
          properties:
            appointments:
              type: array
              items:
                $ref: "#/components/schemas/Appointment"
    Appointment:
      type: object
      required: [id, patientId, patientName, medicalRecordNumber, departmentId, departmentName, doctorId, doctorName, date, time, status, type, createdAt]
      properties:
        id: { type: number }
        patientId: { type: number }
        patientName: { type: string }
        medicalRecordNumber: { type: string }
        departmentId: { type: number }
        departmentName: { type: string }
        doctorId: { type: number }
        doctorName: { type: string }
        date: { type: string, format: date }
        time: { type: string }
        status:
          type: string
          enum: [scheduled, confirmed, waiting, in_progress, completed, cancelled, no_show]
        type:
          type: string
          enum: [first_visit, follow_up, return]
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    AppointmentInput:
      type: object
      required: [patientId, departmentId, doctorId, date, time, type]
      properties:
        patientId: { type: number }
        departmentId: { type: number }
        doctorId: { type: number }
        date: { type: string, format: date }
        time: { type: string }
        type:
          type: string
          enum: [first_visit, follow_up, return]
        notes: { type: string }
    AppointmentUpdate:
      type: object
      properties:
        date: { type: string, format: date }
        time: { type: string }
        doctorId: { type: number }
        status:
          type: string
          enum: [scheduled, confirmed, waiting, in_progress, completed, cancelled, no_show]
        notes: { type: string }
    DashboardSummary:
      type: object
      required: [date, totalAppointments, completedAppointments, waitingPatients, inProgressAppointments, cancelledAppointments, departments, nextAppointment]
      properties:
        date: { type: string, format: date }
        totalAppointments: { type: number }
        completedAppointments: { type: number }
        waitingPatients: { type: number }
        inProgressAppointments: { type: number }
        cancelledAppointments: { type: number }
        averageWaitMinutes: { type: number }
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentSummary"
        nextAppointment:
          anyOf:
            - $ref: "#/components/schemas/Appointment"
            - type: "null"
    DepartmentSummary:
      type: object
      required: [departmentId, departmentName, total, completed, waiting]
      properties:
        departmentId: { type: number }
        departmentName: { type: string }
        total: { type: number }
        completed: { type: number }
        waiting: { type: number }
    ActivityItem:
      type: object
      required: [id, type, message, timestamp]
      properties:
        id: { type: number }
        type:
          type: string
          enum: [appointment_created, patient_registered, status_updated, appointment_completed]
        message: { type: string }
        actor: { type: ["string", "null"] }
        timestamp: { type: string, format: date-time }